backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.6.0
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
alembic==1.10.3
fastapi-limiter==0.1.5
redis==4.5.4
bleach==6.0.0
fastapi-csrf-protect==0.2.2
loguru==0.7.0
tenacity==8.2.2
argon2-cffi==21.3.0
httpx==0.24.0
pytest==7.3.1
pytest-cov==4.1.0
safety==2.3.5
bandit==1.7.5
email-validator==2.0.0
